・モデルファイルの設置場所をapp/Models/⇨app/に修正 ・各メソッドの単数形or複数形にするか再検討し、修正
