+++ app/Follow.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Follow extends Model
+{
+    public function user()
+    {
+        return $this->belongsTo('App\User');
+    }
+}
