--- app/Models/Notification.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Notification extends Model
-{
-    public function users()
-    {
-        return $this->belongsTo('App\Models\User');
-    }
-
-    public function items()
-    {
-        return $this->hasMany('App\Models\Item');
-    }
-}
