+++ app/Notification.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Notification extends Model
+{
+    public function user()
+    {
+        return $this->belongsTo('App\User');
+    }
+
+    public function items()
+    {
+        return $this->hasMany('App\Item');
+    }
+}
